+++ src/pages/Login/Login.container.js
@@ -1,0 +1,33 @@
+import { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { login } from '../../apis/authService';
+import LoginUI from './Login.presenter';
+
+export default function Login() {
+  const [username, setUsername] = useState('');
+  const [password, setPassword] = useState('');
+  const navigate = useNavigate();
+
+  const handleSubmitLogin = async (e) => {
+    e.preventDefault();
+    try {
+      const { accessToken } = await login(username, password);
+      console.log('Access Token: ', accessToken); // Debug log
+      navigate('/');
+    } catch (error) {
+      // eslint-disable-next-line
+      alert('입력 정보를 확인해주세요.');
+      console.error('Login failed', error);
+    }
+  };
+
+  return (
+    <LoginUI
+      username={username}
+      setUsername={setUsername}
+      password={password}
+      setPassword={setPassword}
+      handleSubmitLogin={handleSubmitLogin}
+    />
+  );
+}
